Add option to show measure number interval in mmrests
the MMRestRange is a new class derived from MeasureNumber.
The logic is handled by layoutMeasureNumber, as if it was an usual measurenumber.

MMRestRange may not be styled like MeasureNumber, so a new set of textstyles were added.
There is 3 possible bracketing around the ranges: parentheses, brackets, or none (the default).