icons.cpp: allow more voices than 4.
The current implementation did not allow more than 4 voices, as it relied on a static array of numbers.
Moving it to the simple for loop permits more than 4 voices.

It still does not compile with more than four voices, but it is a step towards.

HOWEVER, since the number of icons in the ICONS enum cannot change, there can only be 4 icons accessible via the icons*.
However again, these icons (voice1_ICON, voice2_ICON, voice3_ICON, voice4_ICON) are not used anywhere in the codebase, so it should not be a problem.